refactoring of models, prepare major release changes
