+++ src/endpoints/zeromq/mod.rs
@@ -1,11 +1,13 @@
 //! ZeroMQ endpoints. Two interchangeable backends share the framing/format
 //! `codec`:
 //!
-//! * [`zmq`] — the default backend on the `zeromq` crate (zmq.rs). Enabled by
-//!   the `zeromq` feature.
-//! * [`omq`] — a PoC backend on `omq-tokio` (omq.rs), PUSH/PULL + PUB/SUB only.
-//!   Enabled by the opt-in `zeromq-omq` feature (MSRV 1.93 / edition 2024, so it
-//!   is kept out of `full`/`portable`).
+//! * [`zmq`] — the `zeromq` crate (zmq.rs), behind the `zeromq` feature.
+//! * [`omq`] — `omq-tokio` (omq.rs), behind the `zeromq-omq` feature (MSRV 1.93 /
+//!   edition 2024, so it is in `full` but not `portable`). Faster on the
+//!   per-message `raw`/`raw_framed` path.
+//!
+//! `backend: try_omq` (the default) picks omq when it is compiled in and falls
+//! back to zmq otherwise; naming a backend explicitly makes it a requirement.
 //!
 //! The backend types are re-exported here so callers keep using
 //! `endpoints::zeromq::{ZeroMqConsumer, ZeroMqPublisher, ...}`.
@@ -26,12 +28,39 @@ use crate::models::{ZeroMqBackend, ZeroMqConfig};
 use crate::traits::{MessageConsumer, MessagePublisher};
 use anyhow::Result;
 
+/// A backend actually to build with — `try_omq` is already resolved away.
+#[derive(Debug, PartialEq, Eq)]
+enum Resolved {
+    Zmq,
+    Omq,
+}
+
+/// Resolve the configured backend to a concrete one.
+///
+/// `try_omq` (the default) prefers omq and falls back to zmq when omq isn't
+/// compiled in, so the same config runs on a build without `zeromq-omq`. An
+/// explicitly named backend is a hard requirement and is kept as-is, so a
+/// missing feature stays a startup error instead of silently changing backend.
+fn resolve_backend(backend: &ZeroMqBackend) -> Resolved {
+    match backend {
+        ZeroMqBackend::Zmq => Resolved::Zmq,
+        ZeroMqBackend::Omq => Resolved::Omq,
+        ZeroMqBackend::TryOmq => {
+            if cfg!(feature = "zeromq-omq") {
+                Resolved::Omq
+            } else {
+                Resolved::Zmq
+            }
+        }
+    }
+}
+
 /// Build a ZeroMQ consumer for the configured backend (`zmq` or `omq`). Each
-/// backend is behind its own build feature; requesting one that wasn't compiled
-/// in is a clear config error rather than a silent fallback.
+/// backend is behind its own build feature; explicitly requesting one that
+/// wasn't compiled in is a clear config error rather than a silent fallback.
 pub(crate) async fn create_consumer(cfg: &ZeroMqConfig) -> Result<Box<dyn MessageConsumer>> {
-    match cfg.backend {
-        ZeroMqBackend::Zmq => {
+    match resolve_backend(&cfg.backend) {
+        Resolved::Zmq => {
             #[cfg(feature = "zeromq")]
             return Ok(Box::new(zmq::ZeroMqConsumer::new(cfg).await?) as Box<dyn MessageConsumer>);
             #[cfg(not(feature = "zeromq"))]
@@ -39,7 +68,7 @@ pub(crate) async fn create_consumer(cfg: &ZeroMqConfig) -> Result<Box<dyn Messag
                 "ZeroMQ backend 'zmq' requires the `zeromq` build feature"
             ));
         }
-        ZeroMqBackend::Omq => {
+        Resolved::Omq => {
             #[cfg(feature = "zeromq-omq")]
             return Ok(
                 Box::new(omq::ZeroMqOmqConsumer::new(cfg).await?) as Box<dyn MessageConsumer>
@@ -54,8 +83,8 @@ pub(crate) async fn create_consumer(cfg: &ZeroMqConfig) -> Result<Box<dyn Messag
 
 /// Publisher counterpart to [`create_consumer`].
 pub(crate) async fn create_publisher(cfg: &ZeroMqConfig) -> Result<Box<dyn MessagePublisher>> {
-    match cfg.backend {
-        ZeroMqBackend::Zmq => {
+    match resolve_backend(&cfg.backend) {
+        Resolved::Zmq => {
             #[cfg(feature = "zeromq")]
             return Ok(Box::new(zmq::ZeroMqPublisher::new(cfg).await?) as Box<dyn MessagePublisher>);
             #[cfg(not(feature = "zeromq"))]
@@ -63,7 +92,7 @@ pub(crate) async fn create_publisher(cfg: &ZeroMqConfig) -> Result<Box<dyn Messa
                 "ZeroMQ backend 'zmq' requires the `zeromq` build feature"
             ));
         }
-        ZeroMqBackend::Omq => {
+        Resolved::Omq => {
             #[cfg(feature = "zeromq-omq")]
             return Ok(
                 Box::new(omq::ZeroMqOmqPublisher::new(cfg).await?) as Box<dyn MessagePublisher>
