+++ tests/integration_test.rs
@@ -495,6 +495,16 @@ async fn test_mongodb_cdc_survives_idle_resume_refresh() {
     }
 }
 
+/// Regression: capture_all must surface an empty batch after its snapshot stays quiet in drain mode.
+#[cfg(feature = "mongodb")]
+#[tokio::test(flavor = "multi_thread")]
+#[ignore = "requires docker compose (mongodb replica set)"]
+async fn test_mongodb_capture_all_exits_on_empty() {
+    if should_run("mongodb_cdc") || should_run("mongodb") {
+        integration::mongodb::test_mongodb_capture_all_exits_on_empty().await;
+    }
+}
+
 #[tokio::test(flavor = "multi_thread")]
 #[ignore = "requires docker compose, takes long time to run"]
 async fn test_all_performance_pipeline() {
